refactor: move entitlements to gamedata service
